+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 ## Changelog
 
-### Newest version: 1.5.1
+### Newest version: 2.0.2
+
+* Fix a bug when titles are always shown 
+
+### 2.0.1
+
+* Fix a crash with `setCurrentItem(int position, boolean useCallback)`
+
+### 2.0.0
+
+* **BREAKING!** 3 states for titles: `SHOW_WHEN_ACTIVE`, `ALWAYS_SHOW` & `ALWAYS_HIDE` (PR #140)
+* Color under the navigation bar (PR #166)
+* Fix CoordinatorLayout with FloatingActionButton: use `manageFloatingActionButtonBehavior`
+
+### 1.5.1
 
 * Fixed ripple effect bug (API 21+)